requirements.txt
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
jieba==0.42.1
py2neo==2021.2.3
PyMySQL==1.0.2
django-cors-headers==3.11.0
beautifulsoup4==4.10.0
paddlepaddle-tiny==1.6.1
lxml==4.8.0
transformers==4.17.0
python-docx==0.8.11
spacy==3.2.4
pip
setuptools
wheel
cryptography==36.0.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.0
psutil==5.9.0
torch==1.12.1